requirements.txt
python-docx==1.1.0
unstructured==0.11.8
pypdf==3.17.4
streamlit==1.29.0
requests==2.31.0
python-dotenv==1.0.0
pandas==2.1.4
numpy==1.26.2
# selenium==4.16.0  # Commented out - not needed for basic version
# webdriver-manager==4.0.1  # Commented out - not needed for basic version
# docx2pdf==0.1.8  # Commented out - not needed for basic version